+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 3.4.1
 * Added Better Haste / Mining Fatigue
   * Removes the attack speed changes from these effects
 * Halved Ancient Debris hardness